// src/app/types/character/about/components/character.attributes.model.ts
export interface CharacterAttributes {             // Атрибуты
  strength: number;        // Сила
  dexterity: number;       // Ловкость
  constitution: number;    // Выносливость
  intelligence: number;    // Интеллект
  wisdom: number;          // Мудрость
  charisma: number;        // Харизма
}